requirements.txt
pandas>=2.0.0
numpy>=1.24.0
tushare>=1.2.89
akshare>=1.10.0
ta-lib>=0.4.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
jinja2>=3.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
